@@ -29,347 +29,6 @@ import (
 )
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationPromotionRun) DeepCopyInto(out *ApplicationPromotionRun) {
-	*out = *in
-	out.TypeMeta = in.TypeMeta
-	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
-	out.Spec = in.Spec
-	in.Status.DeepCopyInto(&out.Status)
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationPromotionRun.
-func (in *ApplicationPromotionRun) DeepCopy() *ApplicationPromotionRun {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationPromotionRun)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
-func (in *ApplicationPromotionRun) DeepCopyObject() runtime.Object {
-	if c := in.DeepCopy(); c != nil {
-		return c
-	}
-	return nil
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationPromotionRunList) DeepCopyInto(out *ApplicationPromotionRunList) {
-	*out = *in
-	out.TypeMeta = in.TypeMeta
-	in.ListMeta.DeepCopyInto(&out.ListMeta)
-	if in.Items != nil {
-		in, out := &in.Items, &out.Items
-		*out = make([]ApplicationPromotionRun, len(*in))
-		for i := range *in {
-			(*in)[i].DeepCopyInto(&(*out)[i])
-		}
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationPromotionRunList.
-func (in *ApplicationPromotionRunList) DeepCopy() *ApplicationPromotionRunList {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationPromotionRunList)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
-func (in *ApplicationPromotionRunList) DeepCopyObject() runtime.Object {
-	if c := in.DeepCopy(); c != nil {
-		return c
-	}
-	return nil
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationPromotionRunSpec) DeepCopyInto(out *ApplicationPromotionRunSpec) {
-	*out = *in
-	out.ManualPromotion = in.ManualPromotion
-	out.AutomatedPromotion = in.AutomatedPromotion
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationPromotionRunSpec.
-func (in *ApplicationPromotionRunSpec) DeepCopy() *ApplicationPromotionRunSpec {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationPromotionRunSpec)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationPromotionRunStatus) DeepCopyInto(out *ApplicationPromotionRunStatus) {
-	*out = *in
-	if in.EnvironmentStatus != nil {
-		in, out := &in.EnvironmentStatus, &out.EnvironmentStatus
-		*out = make([]PromotionRunEnvironmentStatus, len(*in))
-		copy(*out, *in)
-	}
-	if in.ActiveBindings != nil {
-		in, out := &in.ActiveBindings, &out.ActiveBindings
-		*out = make([]string, len(*in))
-		copy(*out, *in)
-	}
-	in.PromotionStartTime.DeepCopyInto(&out.PromotionStartTime)
-	if in.Conditions != nil {
-		in, out := &in.Conditions, &out.Conditions
-		*out = make([]PromotionRunCondition, len(*in))
-		for i := range *in {
-			(*in)[i].DeepCopyInto(&(*out)[i])
-		}
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationPromotionRunStatus.
-func (in *ApplicationPromotionRunStatus) DeepCopy() *ApplicationPromotionRunStatus {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationPromotionRunStatus)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationSnapshot) DeepCopyInto(out *ApplicationSnapshot) {
-	*out = *in
-	out.TypeMeta = in.TypeMeta
-	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
-	in.Spec.DeepCopyInto(&out.Spec)
-	in.Status.DeepCopyInto(&out.Status)
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationSnapshot.
-func (in *ApplicationSnapshot) DeepCopy() *ApplicationSnapshot {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationSnapshot)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
-func (in *ApplicationSnapshot) DeepCopyObject() runtime.Object {
-	if c := in.DeepCopy(); c != nil {
-		return c
-	}
-	return nil
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationSnapshotComponent) DeepCopyInto(out *ApplicationSnapshotComponent) {
-	*out = *in
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationSnapshotComponent.
-func (in *ApplicationSnapshotComponent) DeepCopy() *ApplicationSnapshotComponent {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationSnapshotComponent)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationSnapshotEnvironmentBinding) DeepCopyInto(out *ApplicationSnapshotEnvironmentBinding) {
-	*out = *in
-	out.TypeMeta = in.TypeMeta
-	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
-	in.Spec.DeepCopyInto(&out.Spec)
-	in.Status.DeepCopyInto(&out.Status)
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationSnapshotEnvironmentBinding.
-func (in *ApplicationSnapshotEnvironmentBinding) DeepCopy() *ApplicationSnapshotEnvironmentBinding {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationSnapshotEnvironmentBinding)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
-func (in *ApplicationSnapshotEnvironmentBinding) DeepCopyObject() runtime.Object {
-	if c := in.DeepCopy(); c != nil {
-		return c
-	}
-	return nil
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationSnapshotEnvironmentBindingList) DeepCopyInto(out *ApplicationSnapshotEnvironmentBindingList) {
-	*out = *in
-	out.TypeMeta = in.TypeMeta
-	in.ListMeta.DeepCopyInto(&out.ListMeta)
-	if in.Items != nil {
-		in, out := &in.Items, &out.Items
-		*out = make([]ApplicationSnapshotEnvironmentBinding, len(*in))
-		for i := range *in {
-			(*in)[i].DeepCopyInto(&(*out)[i])
-		}
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationSnapshotEnvironmentBindingList.
-func (in *ApplicationSnapshotEnvironmentBindingList) DeepCopy() *ApplicationSnapshotEnvironmentBindingList {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationSnapshotEnvironmentBindingList)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
-func (in *ApplicationSnapshotEnvironmentBindingList) DeepCopyObject() runtime.Object {
-	if c := in.DeepCopy(); c != nil {
-		return c
-	}
-	return nil
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationSnapshotEnvironmentBindingSpec) DeepCopyInto(out *ApplicationSnapshotEnvironmentBindingSpec) {
-	*out = *in
-	if in.Components != nil {
-		in, out := &in.Components, &out.Components
-		*out = make([]BindingComponent, len(*in))
-		for i := range *in {
-			(*in)[i].DeepCopyInto(&(*out)[i])
-		}
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationSnapshotEnvironmentBindingSpec.
-func (in *ApplicationSnapshotEnvironmentBindingSpec) DeepCopy() *ApplicationSnapshotEnvironmentBindingSpec {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationSnapshotEnvironmentBindingSpec)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationSnapshotEnvironmentBindingStatus) DeepCopyInto(out *ApplicationSnapshotEnvironmentBindingStatus) {
-	*out = *in
-	if in.GitOpsDeployments != nil {
-		in, out := &in.GitOpsDeployments, &out.GitOpsDeployments
-		*out = make([]BindingStatusGitOpsDeployment, len(*in))
-		copy(*out, *in)
-	}
-	if in.Components != nil {
-		in, out := &in.Components, &out.Components
-		*out = make([]ComponentStatus, len(*in))
-		for i := range *in {
-			(*in)[i].DeepCopyInto(&(*out)[i])
-		}
-	}
-	if in.GitOpsRepoConditions != nil {
-		in, out := &in.GitOpsRepoConditions, &out.GitOpsRepoConditions
-		*out = make([]metav1.Condition, len(*in))
-		for i := range *in {
-			(*in)[i].DeepCopyInto(&(*out)[i])
-		}
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationSnapshotEnvironmentBindingStatus.
-func (in *ApplicationSnapshotEnvironmentBindingStatus) DeepCopy() *ApplicationSnapshotEnvironmentBindingStatus {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationSnapshotEnvironmentBindingStatus)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationSnapshotList) DeepCopyInto(out *ApplicationSnapshotList) {
-	*out = *in
-	out.TypeMeta = in.TypeMeta
-	in.ListMeta.DeepCopyInto(&out.ListMeta)
-	if in.Items != nil {
-		in, out := &in.Items, &out.Items
-		*out = make([]ApplicationSnapshot, len(*in))
-		for i := range *in {
-			(*in)[i].DeepCopyInto(&(*out)[i])
-		}
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationSnapshotList.
-func (in *ApplicationSnapshotList) DeepCopy() *ApplicationSnapshotList {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationSnapshotList)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
-func (in *ApplicationSnapshotList) DeepCopyObject() runtime.Object {
-	if c := in.DeepCopy(); c != nil {
-		return c
-	}
-	return nil
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationSnapshotSpec) DeepCopyInto(out *ApplicationSnapshotSpec) {
-	*out = *in
-	if in.Components != nil {
-		in, out := &in.Components, &out.Components
-		*out = make([]ApplicationSnapshotComponent, len(*in))
-		copy(*out, *in)
-	}
-	in.Artifacts.DeepCopyInto(&out.Artifacts)
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationSnapshotSpec.
-func (in *ApplicationSnapshotSpec) DeepCopy() *ApplicationSnapshotSpec {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationSnapshotSpec)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ApplicationSnapshotStatus) DeepCopyInto(out *ApplicationSnapshotStatus) {
-	*out = *in
-	if in.Conditions != nil {
-		in, out := &in.Conditions, &out.Conditions
-		*out = make([]metav1.Condition, len(*in))
-		for i := range *in {
-			(*in)[i].DeepCopyInto(&(*out)[i])
-		}
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ApplicationSnapshotStatus.
-func (in *ApplicationSnapshotStatus) DeepCopy() *ApplicationSnapshotStatus {
-	if in == nil {
-		return nil
-	}
-	out := new(ApplicationSnapshotStatus)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *AutomatedPromotionConfiguration) DeepCopyInto(out *AutomatedPromotionConfiguration) {
 	*out = *in
 }
@@ -642,6 +301,33 @@ func (in *ManualPromotionConfiguration) DeepCopy() *ManualPromotionConfiguration
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *PromotionRun) DeepCopyInto(out *PromotionRun) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
+	out.Spec = in.Spec
+	in.Status.DeepCopyInto(&out.Status)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PromotionRun.
+func (in *PromotionRun) DeepCopy() *PromotionRun {
+	if in == nil {
+		return nil
+	}
+	out := new(PromotionRun)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *PromotionRun) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PromotionRunCondition) DeepCopyInto(out *PromotionRunCondition) {
 	*out = *in
 	in.LastProbeTime.DeepCopyInto(&out.LastProbeTime)
@@ -677,6 +363,115 @@ func (in *PromotionRunEnvironmentStatus) DeepCopy() *PromotionRunEnvironmentStat
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *PromotionRunList) DeepCopyInto(out *PromotionRunList) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ListMeta.DeepCopyInto(&out.ListMeta)
+	if in.Items != nil {
+		in, out := &in.Items, &out.Items
+		*out = make([]PromotionRun, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PromotionRunList.
+func (in *PromotionRunList) DeepCopy() *PromotionRunList {
+	if in == nil {
+		return nil
+	}
+	out := new(PromotionRunList)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *PromotionRunList) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *PromotionRunSpec) DeepCopyInto(out *PromotionRunSpec) {
+	*out = *in
+	out.ManualPromotion = in.ManualPromotion
+	out.AutomatedPromotion = in.AutomatedPromotion
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PromotionRunSpec.
+func (in *PromotionRunSpec) DeepCopy() *PromotionRunSpec {
+	if in == nil {
+		return nil
+	}
+	out := new(PromotionRunSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *PromotionRunStatus) DeepCopyInto(out *PromotionRunStatus) {
+	*out = *in
+	if in.EnvironmentStatus != nil {
+		in, out := &in.EnvironmentStatus, &out.EnvironmentStatus
+		*out = make([]PromotionRunEnvironmentStatus, len(*in))
+		copy(*out, *in)
+	}
+	if in.ActiveBindings != nil {
+		in, out := &in.ActiveBindings, &out.ActiveBindings
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	in.PromotionStartTime.DeepCopyInto(&out.PromotionStartTime)
+	if in.Conditions != nil {
+		in, out := &in.Conditions, &out.Conditions
+		*out = make([]PromotionRunCondition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PromotionRunStatus.
+func (in *PromotionRunStatus) DeepCopy() *PromotionRunStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(PromotionRunStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *Snapshot) DeepCopyInto(out *Snapshot) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
+	in.Spec.DeepCopyInto(&out.Spec)
+	in.Status.DeepCopyInto(&out.Status)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new Snapshot.
+func (in *Snapshot) DeepCopy() *Snapshot {
+	if in == nil {
+		return nil
+	}
+	out := new(Snapshot)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *Snapshot) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *SnapshotArtifacts) DeepCopyInto(out *SnapshotArtifacts) {
 	*out = *in
 	if in.UnstableFields != nil {
@@ -697,9 +492,222 @@ func (in *SnapshotArtifacts) DeepCopy() *SnapshotArtifacts {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *SnapshotComponent) DeepCopyInto(out *SnapshotComponent) {
+	*out = *in
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new SnapshotComponent.
+func (in *SnapshotComponent) DeepCopy() *SnapshotComponent {
+	if in == nil {
+		return nil
+	}
+	out := new(SnapshotComponent)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *SnapshotEnvironmentBinding) DeepCopyInto(out *SnapshotEnvironmentBinding) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
+	in.Spec.DeepCopyInto(&out.Spec)
+	in.Status.DeepCopyInto(&out.Status)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new SnapshotEnvironmentBinding.
+func (in *SnapshotEnvironmentBinding) DeepCopy() *SnapshotEnvironmentBinding {
+	if in == nil {
+		return nil
+	}
+	out := new(SnapshotEnvironmentBinding)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *SnapshotEnvironmentBinding) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *SnapshotEnvironmentBindingList) DeepCopyInto(out *SnapshotEnvironmentBindingList) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ListMeta.DeepCopyInto(&out.ListMeta)
+	if in.Items != nil {
+		in, out := &in.Items, &out.Items
+		*out = make([]SnapshotEnvironmentBinding, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new SnapshotEnvironmentBindingList.
+func (in *SnapshotEnvironmentBindingList) DeepCopy() *SnapshotEnvironmentBindingList {
+	if in == nil {
+		return nil
+	}
+	out := new(SnapshotEnvironmentBindingList)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *SnapshotEnvironmentBindingList) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *SnapshotEnvironmentBindingSpec) DeepCopyInto(out *SnapshotEnvironmentBindingSpec) {
+	*out = *in
+	if in.Components != nil {
+		in, out := &in.Components, &out.Components
+		*out = make([]BindingComponent, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new SnapshotEnvironmentBindingSpec.
+func (in *SnapshotEnvironmentBindingSpec) DeepCopy() *SnapshotEnvironmentBindingSpec {
+	if in == nil {
+		return nil
+	}
+	out := new(SnapshotEnvironmentBindingSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *SnapshotEnvironmentBindingStatus) DeepCopyInto(out *SnapshotEnvironmentBindingStatus) {
+	*out = *in
+	if in.GitOpsDeployments != nil {
+		in, out := &in.GitOpsDeployments, &out.GitOpsDeployments
+		*out = make([]BindingStatusGitOpsDeployment, len(*in))
+		copy(*out, *in)
+	}
+	if in.Components != nil {
+		in, out := &in.Components, &out.Components
+		*out = make([]ComponentStatus, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+	if in.GitOpsRepoConditions != nil {
+		in, out := &in.GitOpsRepoConditions, &out.GitOpsRepoConditions
+		*out = make([]metav1.Condition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+	if in.BindingConditions != nil {
+		in, out := &in.BindingConditions, &out.BindingConditions
+		*out = make([]metav1.Condition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new SnapshotEnvironmentBindingStatus.
+func (in *SnapshotEnvironmentBindingStatus) DeepCopy() *SnapshotEnvironmentBindingStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(SnapshotEnvironmentBindingStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *SnapshotList) DeepCopyInto(out *SnapshotList) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ListMeta.DeepCopyInto(&out.ListMeta)
+	if in.Items != nil {
+		in, out := &in.Items, &out.Items
+		*out = make([]Snapshot, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new SnapshotList.
+func (in *SnapshotList) DeepCopy() *SnapshotList {
+	if in == nil {
+		return nil
+	}
+	out := new(SnapshotList)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *SnapshotList) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *SnapshotSpec) DeepCopyInto(out *SnapshotSpec) {
+	*out = *in
+	if in.Components != nil {
+		in, out := &in.Components, &out.Components
+		*out = make([]SnapshotComponent, len(*in))
+		copy(*out, *in)
+	}
+	in.Artifacts.DeepCopyInto(&out.Artifacts)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new SnapshotSpec.
+func (in *SnapshotSpec) DeepCopy() *SnapshotSpec {
+	if in == nil {
+		return nil
+	}
+	out := new(SnapshotSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *SnapshotStatus) DeepCopyInto(out *SnapshotStatus) {
+	*out = *in
+	if in.Conditions != nil {
+		in, out := &in.Conditions, &out.Conditions
+		*out = make([]metav1.Condition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new SnapshotStatus.
+func (in *SnapshotStatus) DeepCopy() *SnapshotStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(SnapshotStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *UnstableEnvironmentConfiguration) DeepCopyInto(out *UnstableEnvironmentConfiguration) {
 	*out = *in
 	out.KubernetesClusterCredentials = in.KubernetesClusterCredentials
+	out.SubWorkspace = in.SubWorkspace
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new UnstableEnvironmentConfiguration.
