@@ -1,6 +1,8 @@
 package application_event_loop
 
 import (
+	"context"
+
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/redhat-appstudio/managed-gitops/backend/eventloop/eventlooptypes"
@@ -19,17 +21,16 @@ var _ = Describe("ApplicationEventLoop Test", func() {
 				mockChannel: make(chan *eventlooptypes.EventLoopEvent),
 			}
 
-			inputChan := startApplicationEventQueueLoopWithFactory("", "", nil, &mockApplicationEventLoopRunnerFactory)
+			inputChan := startApplicationEventQueueLoopWithFactory(context.Background(), "", "", "", nil, &mockApplicationEventLoopRunnerFactory)
 
 			inputChan <- eventlooptypes.EventLoopMessage{
 				MessageType: eventlooptypes.ApplicationEventLoopMessageType_Event,
 				Event: &eventlooptypes.EventLoopEvent{
-					EventType:               eventlooptypes.DeploymentModified,
-					Request:                 reconcile.Request{NamespacedName: types.NamespacedName{Namespace: "", Name: ""}},
-					Client:                  nil,
-					ReqResource:             eventlooptypes.GitOpsDeploymentTypeName,
-					AssociatedGitopsDeplUID: "",
-					WorkspaceID:             "",
+					EventType:   eventlooptypes.DeploymentModified,
+					Request:     reconcile.Request{NamespacedName: types.NamespacedName{Namespace: "", Name: ""}},
+					Client:      nil,
+					ReqResource: eventlooptypes.GitOpsDeploymentTypeName,
+					WorkspaceID: "",
 				},
 				ShutdownSignalled: false,
 			}
@@ -39,7 +40,6 @@ var _ = Describe("ApplicationEventLoop Test", func() {
 			Expect(outputEvent.EventType).To(Equal(outputEvent.EventType))
 			Expect(outputEvent.Request).To(Equal(outputEvent.Request))
 			Expect(outputEvent.ReqResource).To(Equal(outputEvent.ReqResource))
-			Expect(outputEvent.AssociatedGitopsDeplUID).To(Equal(outputEvent.AssociatedGitopsDeplUID))
 			Expect(outputEvent.WorkspaceID).To(Equal(outputEvent.WorkspaceID))
 
 		})
@@ -55,8 +55,8 @@ type mockApplicationEventLoopRunnerFactory struct {
 var _ applicationEventRunnerFactory = &mockApplicationEventLoopRunnerFactory{}
 
 func (fact *mockApplicationEventLoopRunnerFactory) createNewApplicationEventLoopRunner(informWorkCompleteChan chan eventlooptypes.EventLoopMessage,
-	sharedResourceEventLoop *shared_resource_loop.SharedResourceEventLoop,
-	gitopsDeplUID string, workspaceID string, debugContext string) chan *eventlooptypes.EventLoopEvent {
+	sharedResourceEventLoop *shared_resource_loop.SharedResourceEventLoop, gitopsDeplName string, gitopsDeplNamespace string,
+	workspaceID string, debugContext string) chan *eventlooptypes.EventLoopEvent {
 
 	return fact.mockChannel
 
