@@ -7,6 +7,7 @@ import (
 	managedgitopsv1alpha1 "github.com/redhat-appstudio/managed-gitops/backend-shared/apis/managed-gitops/v1alpha1"
 	db "github.com/redhat-appstudio/managed-gitops/backend-shared/config/db"
 	dbutil "github.com/redhat-appstudio/managed-gitops/backend-shared/config/db/util"
+	"github.com/redhat-appstudio/managed-gitops/backend-shared/util/gitopserrors"
 	"github.com/redhat-appstudio/managed-gitops/backend-shared/util/operations"
 	"github.com/redhat-appstudio/managed-gitops/backend/eventloop/eventlooptypes"
 
@@ -20,18 +21,55 @@ import (
 
 // This file is responsible for processing events related to GitOpsDeploymentSyncRun CR.
 
-func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRunModified(ctx context.Context, dbQueries db.ApplicationScopedQueries) (bool, error) {
+func (action *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRunModified(ctx context.Context, dbQueries db.ApplicationScopedQueries) (bool, error) {
+
+	// Handle all GitOpsDeployment related events
+	signalledShutdown, err := action.applicationEventRunner_handleSyncRunModifiedInternal(ctx, dbQueries)
+
+	// TODO: GITOPSRVCE-44: Get the SyncRunModified object from k8s, so we can update it if necessary
+
+	// TODO: GITOPSRVCE-44: Add the ErrorOccurred condition to the syncRunModified
+
+	// TODO: GITOPSRVCE-44: If no user error (just dev error), then output generic error occurred
+
+	if err == nil {
+		return signalledShutdown, nil
+	} else {
+		return signalledShutdown, err.DevError()
+	}
+
+}
+
+func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRunModifiedInternal(ctx context.Context,
+	dbQueries db.ApplicationScopedQueries) (bool, gitopserrors.UserError) {
+
+	// TODO: GITOPSRVCE-166: This function can likely be broken up into a few smaller function, for readability/maintainability.
+
+	const (
+		// TODO: GITOPSRVCE-166: I'm not convinced that these are needed, and can thus be removed.
+		signalledShutdown_true  = true
+		signalledShutdown_false = false
+	)
 
 	log := a.log
 
+	deplName := a.eventResourceName
+	deplNamespace := a.eventResourceNamespace
+
 	namespace := corev1.Namespace{}
 	if err := a.workspaceClient.Get(ctx, types.NamespacedName{Namespace: a.eventResourceNamespace, Name: a.eventResourceNamespace}, &namespace); err != nil {
-		return false, fmt.Errorf("unable to retrieve namespace '%s': %v", a.eventResourceNamespace, err)
+		userError := fmt.Sprintf("unable to retrieve the contents of the namespace '%s' containing the API resource '%s'. Does it exist?",
+			deplNamespace, deplName)
+		devError := fmt.Errorf("unable to retrieve namespace '%s': %v", deplNamespace, err)
+		return signalledShutdown_false, gitopserrors.NewUserDevError(userError, devError)
 	}
 
 	clusterUser, _, err := a.sharedResourceEventLoop.GetOrCreateClusterUserByNamespaceUID(ctx, a.workspaceClient, namespace, log)
 	if err != nil {
-		return false, fmt.Errorf("unable to retrieve cluster user in handleDeploymentModified, '%s': %v", string(namespace.UID), err)
+		userError := "unable to locate managed environment for new application"
+		devError := fmt.Errorf("unable to retrieve cluster user in applicationEventRunner_handleSyncRunModifiedInternal, '%s': %v",
+			string(namespace.UID), err)
+		return signalledShutdown_false, gitopserrors.NewUserDevError(userError, devError)
 	}
 
 	// Retrieve the GitOpsDeploymentSyncRun from the namespace
@@ -45,15 +83,18 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 			if apierr.IsNotFound(err) {
 				syncRunCRExists = false
 			} else {
+				userError := "unable to retrieve the GitOpsDeploymentSyncRun object from the namespace, due to unknown error."
 				log.Error(err, "unable to locate object in handleSyncRunModified", "request", syncRunKey)
-				return false, err
+				return signalledShutdown_false, gitopserrors.NewUserDevError(userError, err)
 			}
 		}
 	}
 
+	// Retrieve the SyncOperation row that corresponds to the SyncRun resource
 	var apiCRToDBList []db.APICRToDatabaseMapping
 	dbEntryExists := false
 	if syncRunCRExists {
+
 		mapping := db.APICRToDatabaseMapping{
 			APIResourceType: db.APICRToDatabaseMapping_ResourceType_GitOpsDeploymentSyncRun,
 			APIResourceUID:  string(syncRunCR.UID),
@@ -66,24 +107,27 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 				dbEntryExists = false
 			} else {
 				// A generic error occured, so just return
+				userError := "unable to retrieve GitOpsDeploymentSyncRun metadata from the internal database, due to an unknown error"
 				log.Error(err, "unable to resource APICRToDatabaseMapping", "uid", string(syncRunCR.UID))
-				return false, err
+				return signalledShutdown_false, gitopserrors.NewUserDevError(userError, err)
 			}
 		} else {
 			// Match found in database
 			apiCRToDBList = append(apiCRToDBList, mapping)
 		}
+
 	} else {
+
 		// The CR no longer exists (it was likely deleted), so instead we retrieve the UID of the SyncRun from
-		// the APICRToDatabaseMapping table, by combination of (name/namespace/workspace).
+		// the APICRToDatabaseMapping table, by combination of (name/namespace/namespace uid).
 
 		if err := dbQueries.ListAPICRToDatabaseMappingByAPINamespaceAndName(ctx, db.APICRToDatabaseMapping_ResourceType_GitOpsDeploymentSyncRun,
 			a.eventResourceName, a.eventResourceNamespace, eventlooptypes.GetWorkspaceIDFromNamespaceID(namespace),
 			db.APICRToDatabaseMapping_DBRelationType_SyncOperation, &apiCRToDBList); err != nil {
-
+			userError := "unable to retrive data related to previous GitOpsDeploymentSyncRun in the namespace, due to an unknown error"
 			log.Error(err, "unable to find API CR to DB Mapping, by API name/namespace/uid",
 				"name", a.eventResourceName, "namespace", a.eventResourceNamespace, "UID", string(namespace.UID))
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewUserDevError(userError, err)
 		}
 
 		if len(apiCRToDBList) == 0 {
@@ -96,19 +140,24 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		}
 	}
 
-	log.Info("workspacerEventLoopRunner_handleSyncRunModified", "syncRunCRExists", syncRunCRExists, "dbEntryExists", dbEntryExists)
+	log.Info("workspaceEventLoopRunner_handleSyncRunModified", "syncRunCRExists", syncRunCRExists, "dbEntryExists", dbEntryExists)
 
 	if !syncRunCRExists && !dbEntryExists {
 		log.Info("neither sync run CR exists, nor db entry, so our work is done.")
 		// if neither exists, our work is done
-		return true, nil
+		return signalledShutdown_true, nil
 	}
 
 	// The applications and gitopsengineinstance pointed to by the gitopsdeployment (if they are non-nil)
 	var application *db.Application
 	var gitopsEngineInstance *db.GitopsEngineInstance
 
+	// TODO: GITOPSRVCE-166: Make sure that the GitOpsDeployment UID doesn't change on us, vs what is in SyncOperation in the database.
+
 	if syncRunCRExists {
+
+		// TODO: GITOPSRVCE-166: It seems like this needs to change: we should retrieve the SyncOperation associated with the GitOpsDeploymentSyncRun, and sanity check it is equal. This can be handled while refactoring this function.
+
 		// Sanity check that the gitopsdeployment resource exists, which is referenced by the syncrun resource
 		gitopsDepl := &managedgitopsv1alpha1.GitOpsDeployment{
 			ObjectMeta: metav1.ObjectMeta{
@@ -125,14 +174,14 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 				err := fmt.Errorf("unable to retrieve gitopsdeployment referenced in syncrun: %v", err)
 				// TODO: GITOPSRVCE-44 - ENHANCEMENT - If the gitopsDepl isn't referenced, update the status of the GitOpsDeplomentSyncRun condition as an error and return
 				// TODO: GITOPSRVCE-44 - ENHANCEMENT - implement status conditions on GitOpsDeploymentSyncRun
+				userError := fmt.Sprintf("Unable to retrieve GitOpsDeployment '%s' referenced by the GitOpsDeploymentSyncRun", gitopsDepl.Name)
 				log.Error(err, "handleSyncRunModified error")
-				return false, err
-
+				return signalledShutdown_false, gitopserrors.NewUserDevError(userError, err)
 			}
 
 			// If there was a generic error in retrieving the key, return it
 			log.Error(err, "unable to retrieve gitopsdeployment referenced in syncrun")
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(fmt.Errorf("SEVERE - All cases should be handled by above if statements"))
 		}
 
 		// The GitopsDepl CR exists, so use the UID of the CR to retrieve the database entry, if possible
@@ -140,20 +189,20 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 
 		if err = dbQueries.GetDeploymentToApplicationMappingByDeplId(ctx, deplToAppMapping); err != nil {
 			log.Error(err, "unable to retrieve deployment to application mapping, on sync run modified", "uid", string(gitopsDepl.UID))
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		application = &db.Application{Application_id: deplToAppMapping.Application_id}
 		if err := dbQueries.GetApplicationById(ctx, application); err != nil {
 			log.Error(err, "unable to retrieve application, on sync run modified", "applicationId", string(deplToAppMapping.Application_id))
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		if gitopsEngineInstance, err = a.sharedResourceEventLoop.GetGitopsEngineInstanceById(ctx, application.Engine_instance_inst_id,
 			a.workspaceClient, namespace, a.log); err != nil {
 
 			log.Error(err, "unable to retrieve gitopsengineinstance, on sync run modified", "instanceId", string(application.Engine_instance_inst_id))
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 	}
@@ -176,7 +225,7 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		if application == nil || gitopsEngineInstance == nil {
 			err := fmt.Errorf("app or engine instance were nil in handleSyncRunModified app: %v, instance: %v", application, gitopsEngineInstance)
 			log.Error(err, "unexpected nil value of required objects")
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		// createdResources is a list of database entries created in this function; if an error occurs, we delete them
@@ -193,7 +242,7 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		if err := dbQueries.CreateSyncOperation(ctx, syncOperation); err != nil {
 			log.Error(err, "unable to create sync operation in database")
 
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 		createdResources = append(createdResources, syncOperation)
 		log.Info("Created a Sync Operation: " + syncOperation.SyncOperation_id)
@@ -214,7 +263,7 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 			// If we were unable to retrieve the client, delete the resources we created in the previous steps
 			dbutil.DisposeApplicationScopedResources(ctx, createdResources, dbQueries, log)
 
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 		log.Info(fmt.Sprintf("Created a ApiCRToDBMapping: (APIResourceType: %s, APIResourceUID: %s, DBRelationType: %s)", newApiCRToDBMapping.APIResourceType, newApiCRToDBMapping.APIResourceUID, newApiCRToDBMapping.DBRelationType))
 		createdResources = append(createdResources, &newApiCRToDBMapping)
@@ -227,7 +276,7 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 			dbutil.DisposeApplicationScopedResources(ctx, createdResources, dbQueries, log)
 
 			// Return the original error
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		dbOperationInput := db.Operation{
@@ -244,17 +293,17 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 			// If we were unable to create the operation, delete the resources we created in the previous steps
 			dbutil.DisposeApplicationScopedResources(ctx, createdResources, dbQueries, log)
 
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		// TODO: GITOPSRVCE-82 - STUB - Remove the 'false' in createOperation above, once cluster agent handling of operation is implemented.
 		log.Info("STUB: Not waiting for create Sync Run operation to complete, in handleNewSyncRunModified")
 
 		if err := operations.CleanupOperation(ctx, *dbOperation, *k8sOperation, dbutil.GetGitOpsEngineSingleInstanceNamespace(), dbQueries, operationClient, log); err != nil {
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
-		return false, nil
+		return signalledShutdown_false, nil
 	}
 
 	if !syncRunCRExists && dbEntryExists {
@@ -266,7 +315,7 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		if len(apiCRToDBList) != 1 {
 			err := fmt.Errorf("SEVERE - Update only supports one operation parameter")
 			log.Error(err, err.Error())
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		// 1) Get the SyncOperation table entry pointed to by the resource
@@ -275,14 +324,14 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		if apiCRToDBMapping.DBRelationType != db.APICRToDatabaseMapping_DBRelationType_SyncOperation {
 			err := fmt.Errorf("SEVERE - db relation type should be syncoperation")
 			log.Error(err, err.Error())
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		syncOperation := db.SyncOperation{SyncOperation_id: apiCRToDBMapping.DBRelationKey}
 
 		if err := dbQueries.GetSyncOperationById(ctx, &syncOperation); err != nil {
 			log.Error(err, "unable to retrieve sync operation by id on deleted", "operationID", syncOperation.SyncOperation_id)
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		// 2) Update the state of the SyncOperation DB table to say that we want to terminate it, if it is runing
@@ -298,7 +347,7 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		operationClient, err := a.getK8sClientForGitOpsEngineInstance(gitopsEngineInstance)
 		if err != nil {
 			log.Error(err, "unable to retrieve gitopsengine instance from handleSyncRunModified, when resource was deleted")
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		waitForOperation := !a.testOnlySkipCreateOperation // if it's for a unit test, we don't wait for the operation
@@ -307,12 +356,12 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		if err != nil {
 			log.Error(err, "could not create operation, when resource was deleted", "namespace", dbutil.GetGitOpsEngineSingleInstanceNamespace())
 
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		// 4) Clean up the operation and database table entries
 		if err := operations.CleanupOperation(ctx, *dbOperation, *k8sOperation, dbutil.GetGitOpsEngineSingleInstanceNamespace(), dbQueries, operationClient, log); err != nil {
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		// TODO: GITOPSRVCE-82 - STUB - need to implement support for sync operation in cluster agent
@@ -320,7 +369,7 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 
 		if _, err := dbQueries.DeleteSyncOperationById(ctx, syncOperation.SyncOperation_id); err != nil {
 			log.Error(err, "could not delete sync operation, when resource was deleted", "namespace", dbutil.GetGitOpsEngineSingleInstanceNamespace())
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		} else {
 			log.Info("Sync Operation deleted with ID: ", syncOperation.SyncOperation_id)
 		}
@@ -343,11 +392,11 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		}
 
 		if allErrors != nil {
-			return false, allErrors
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(allErrors)
 		}
 
 		// Success: the CR no longer exists, and we have completed cleanup, so signal that the goroutine may be terminated.
-		return true, nil
+		return signalledShutdown_true, nil
 	}
 
 	if syncRunCRExists && dbEntryExists {
@@ -356,13 +405,13 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		if syncRunCR == (&managedgitopsv1alpha1.GitOpsDeploymentSyncRun{}) {
 			err := fmt.Errorf("SEVERE - vsync run cr is empty")
 			log.Error(err, err.Error())
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		if len(apiCRToDBList) != 1 {
 			err := fmt.Errorf("SEVERE - Update only supports one operation parameter")
 			log.Error(err, err.Error())
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		// Get the SyncOperation table entry pointed to by the resource
@@ -371,7 +420,7 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		if apiCRToDBMapping.DBRelationType != db.APICRToDatabaseMapping_DBRelationType_SyncOperation {
 			err := fmt.Errorf("SEVERE - db relation type should be syncoperation")
 			log.Error(err, err.Error())
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		syncOperation := db.SyncOperation{SyncOperation_id: apiCRToDBMapping.DBRelationKey}
@@ -379,27 +428,29 @@ func (a *applicationEventLoopRunner_Action) applicationEventRunner_handleSyncRun
 		if err := dbQueries.GetSyncOperationById(ctx, &syncOperation); err != nil {
 
 			log.Error(err, "unable to retrieve sync operation by id on modified", "operationID", syncOperation.SyncOperation_id)
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewDevOnlyError(err)
 		}
 
 		if syncOperation.DeploymentNameField != syncRunCR.Spec.GitopsDeploymentName {
-			err := fmt.Errorf("deployment name field is immutable: changing it from its initial value is not supported")
-			log.Error(err, "deployment name field change is not supported")
-			return false, err
+			userErrorText := "deployment name field is immutable: changing it from its initial value is not supported"
+			err := fmt.Errorf(userErrorText)
+			log.Error(err, userErrorText)
+			return signalledShutdown_false, gitopserrors.NewUserDevError(userErrorText, err)
 		}
 
 		if syncOperation.Revision != syncRunCR.Spec.RevisionID {
-			err := fmt.Errorf("revision change is not supported: changing it from its initial value is not supported")
+			userErrorText := "revision change is not supported: changing it from its initial value is not supported"
+			err := fmt.Errorf(userErrorText)
 			log.Error(err, "revision field change is not supported")
-			return false, err
+			return signalledShutdown_false, gitopserrors.NewUserDevError(userErrorText, err)
 		}
 
 		// TODO: GITOPSRVCE-166 - DEBT - Include test case to check that the various goroutines are terminated when the CR is deleted.
 
-		return false, nil
+		return signalledShutdown_false, nil
 	}
 
-	return false, nil
+	return signalledShutdown_false, nil
 
 }
 
