@@ -3,7 +3,6 @@ package eventloop
 import (
 	"context"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/go-logr/logr"
@@ -125,14 +124,14 @@ func internalStartWorkspaceEventLoopRouter(input chan workspaceEventLoopMessage,
 
 }
 
-// workspaceEventLoopRouter receives all events for the workspace, and passes them to specific goroutine responsible
+// workspaceEventLoopRouter receives all events for the namespace, and passes them to specific goroutine responsible
 // for handling events for individual applications.
-func workspaceEventLoopRouter(input chan workspaceEventLoopMessage, workspaceID string,
+func workspaceEventLoopRouter(input chan workspaceEventLoopMessage, namespaceID string,
 	applEventLoopFactory applicationEventQueueLoopFactory) {
 
 	ctx := context.Background()
 
-	log := log.FromContext(ctx).WithValues("workspaceID", workspaceID).WithName("workspace-event-loop")
+	log := log.FromContext(ctx).WithValues("namespaceID", namespaceID).WithName("workspace-event-loop")
 
 	log.Info("workspaceEventLoopRouter started")
 	defer log.Info("workspaceEventLoopRouter ended.")
@@ -141,16 +140,22 @@ func workspaceEventLoopRouter(input chan workspaceEventLoopMessage, workspaceID 
 
 	workspaceResourceLoop := newWorkspaceResourceLoop(sharedResourceEventLoop, input)
 
-	// orphanedResources: gitops depl name -> (name field of CR -> event depending on it)
+	// orphanedResources:
+	// - key: gitops depl name
+	// - value: map: name field of orphaned syncrun CR -> last event that referenced it
 	orphanedResources := map[string]map[string]eventlooptypes.EventLoopEvent{}
 
-	// applicationMap: gitopsDepl UID -> channel for go routine responsible for handling it
+	// applicationMap: maintains a list of which which GitOpsDeployment resources (1-1 relationship) are handled by which Go Routines
+	// - key: string, of format "(namespace UID)-(namespace name)-(GitOpsDeployment name)"
+	// - value:  channel for the go routine responsible for handling events for this GitOpsDeployment
 	applicationMap := map[string]workspaceEventLoop_applicationEventLoopEntry{}
 	for {
 		wrapperEvent := <-input
 
+		event := (wrapperEvent.payload).(eventlooptypes.EventLoopMessage)
+		ctx = sharedutil.AddKCPClusterToContext(ctx, event.Event.Request.ClusterName)
+
 		if wrapperEvent.messageType == workspaceEventLoopMessageType_Event {
-			event := (wrapperEvent.payload).(eventlooptypes.EventLoopMessage)
 			// First, sanity check the event
 			if event.MessageType == eventlooptypes.ApplicationEventLoopMessageType_WorkComplete {
 				log.Error(nil, "SEVERE: invalid message type received in applicationEventLooRouter")
@@ -160,21 +165,17 @@ func workspaceEventLoopRouter(input chan workspaceEventLoopMessage, workspaceID 
 				log.Error(nil, "SEVERE: event was nil in workspaceEventLooRouter")
 				continue
 			}
-			if strings.TrimSpace(event.Event.AssociatedGitopsDeplUID) == "" {
-				log.Error(nil, "SEVERE: event was nil in workspaceEventLoopRouter")
-				continue
-			}
 
-			log.V(sharedutil.LogLevel_Debug).Info("workspaceEventLoop received event", "event", eventlooptypes.StringEventLoopEvent(event.Event))
+			log.V(sharedutil.LogLevel_Debug).Info("workspaceEventLoop received event",
+				"event", eventlooptypes.StringEventLoopEvent(event.Event))
 
-			// If the event is orphaned (it refers to a gitopsdepl that doesn't exist, add it to our orphaned resources list)
-			if event.Event.AssociatedGitopsDeplUID == eventlooptypes.OrphanedResourceGitopsDeplUID {
-				handleOrphaned(ctx, event, orphanedResources, log)
-				continue
-			}
+			// If it's a GitOpsDeploymentSyncRun:
+			//   - If the SyncRun exists, use the name that was specified
+			//   - If the Syncrun doesn't exist, retrieve from the database
+			//     - If it's not found in the database, it's orphaned.
 
-			// Check GitOpsDeployment that come in: if we get an event for GitOpsDeployment that has an orphaned child resource
-			// depending on it, then unorphan the child resource.
+			// Check GitOpsDeployment resource events that come in: if we get an event for GitOpsDeployment that has an
+			// orphaned child resource depending on it, then unorphan the child resource.
 			if event.Event.ReqResource == eventlooptypes.GitOpsDeploymentTypeName {
 				unorphanResourcesIfPossible(ctx, event, orphanedResources, input, log)
 			}
@@ -194,18 +195,41 @@ func workspaceEventLoopRouter(input chan workspaceEventLoopMessage, workspaceID 
 				continue
 			}
 
-			applicationEntryVal, exists := applicationMap[event.Event.AssociatedGitopsDeplUID]
+			var associatedGitOpsDeploymentName string // name of the GitOpsDeployment (in this namespace) that this resource is associated with
+
+			if event.Event.ReqResource == eventlooptypes.GitOpsDeploymentTypeName {
+				associatedGitOpsDeploymentName = event.Event.Request.Name
+
+			} else if event.Event.ReqResource == eventlooptypes.GitOpsDeploymentSyncRunTypeName {
+
+				associatedGitOpsDeploymentName = checkIfOrphanedGitOpsDeploymentSyncRun(ctx, event, orphanedResources, log)
+
+				if associatedGitOpsDeploymentName == "" {
+					// The SyncRun no longer exists, or an unrecoverable error occurred, so just continue
+					continue
+				}
+			}
+
+			if associatedGitOpsDeploymentName == "" {
+				// Sanity check the name field is non-empty
+				log.Error(nil, "SEVERE: received a resource event that we could not determine a GitOpsDeployment name for", "event", eventlooptypes.StringEventLoopEvent(event.Event))
+				continue
+			}
+
+			mapKey := namespaceID + "-" + event.Event.Request.Namespace + "-" + associatedGitOpsDeploymentName
+
+			applicationEntryVal, exists := applicationMap[mapKey]
 			if !exists {
 				// Start the application event queue go-routine, if it's not already started.
 
 				// Start the application event loop's goroutine
-				inputChan := applEventLoopFactory.startApplicationEventQueueLoop(event.Event.AssociatedGitopsDeplUID,
+				inputChan := applEventLoopFactory.startApplicationEventQueueLoop(ctx, associatedGitOpsDeploymentName, event.Event.Request.Namespace,
 					event.Event.WorkspaceID, sharedResourceEventLoop)
 
 				applicationEntryVal = workspaceEventLoop_applicationEventLoopEntry{
 					input: inputChan,
 				}
-				applicationMap[event.Event.AssociatedGitopsDeplUID] = applicationEntryVal
+				applicationMap[mapKey] = applicationEntryVal
 			}
 
 			// Send the event to the channel/go routine that handles all events for this application/gitopsdepl (non-blocking)
@@ -215,7 +239,6 @@ func workspaceEventLoopRouter(input chan workspaceEventLoopMessage, workspaceID 
 			}
 
 		} else if wrapperEvent.messageType == managedEnvProcessed_Event {
-			event := (wrapperEvent.payload).(eventlooptypes.EventLoopMessage)
 
 			log.V(sharedutil.LogLevel_Debug).Info(fmt.Sprintf("received ManagedEnvironment event, passed event to %d applications",
 				len(applicationMap)))
@@ -248,7 +271,7 @@ type workspaceEventLoop_applicationEventLoopEntry struct {
 //
 // The defaultApplicationEventLoopFactory should be used in all cases, except for when writing mocks for unit tests.
 type applicationEventQueueLoopFactory interface {
-	startApplicationEventQueueLoop(gitopsDeplID string, workspaceID string,
+	startApplicationEventQueueLoop(ctx context.Context, gitopsDeplName string, gitopsDeplNamespace string, workspaceID string,
 		sharedResourceEventLoop *shared_resource_loop.SharedResourceEventLoop) chan eventlooptypes.EventLoopMessage
 }
 
@@ -257,22 +280,29 @@ type defaultApplicationEventLoopFactory struct {
 
 // The default implementation of startApplicationEventQueueLoop is just a simple wrapper around a call to
 // StartApplicationEventQueueLoop
-func (defaultApplicationEventLoopFactory) startApplicationEventQueueLoop(gitopsDeplID string, workspaceID string,
-	sharedResourceEventLoop *shared_resource_loop.SharedResourceEventLoop) chan eventlooptypes.EventLoopMessage {
+func (defaultApplicationEventLoopFactory) startApplicationEventQueueLoop(ctx context.Context, gitopsDeplName string, gitopsDeplNamespace string,
+	workspaceID string, sharedResourceEventLoop *shared_resource_loop.SharedResourceEventLoop) chan eventlooptypes.EventLoopMessage {
 
-	res := application_event_loop.StartApplicationEventQueueLoop(gitopsDeplID, workspaceID, sharedResourceEventLoop)
+	res := application_event_loop.StartApplicationEventQueueLoop(ctx, gitopsDeplName, gitopsDeplNamespace, workspaceID, sharedResourceEventLoop)
 
 	return res
 }
 
 var _ applicationEventQueueLoopFactory = defaultApplicationEventLoopFactory{}
 
-// Add a resource to orphaned list (why? because it is a gitopsdeplsyncrun that refers to a gitopsdepl that doesn't exist)
-// See https://docs.google.com/document/d/1e1UwCbwK-Ew5ODWedqp_jZmhiZzYWaxEvIL-tqebMzo/edit#heading=h.8tiycl1h7rns for details.
-func handleOrphaned(ctx context.Context, event eventlooptypes.EventLoopMessage, orphanedResources map[string]map[string]eventlooptypes.EventLoopEvent, log logr.Logger) {
+// Processes events related to GitOpsDeploymentSyncRuns: determines whether the referenced GitOpsDeployment exists.
+// - If the GitOpsDepl exists, return the name
+// - Otherwise, add the SyncRun resource to orphaned list (why? because it is a gitopsdeplsyncrun that refers to a gitopsdepl that doesn't exist)
+//
+// If the resource is orphaned, or an error occurred, "" is returned.
+//
+//   See https://docs.google.com/document/d/1e1UwCbwK-Ew5ODWedqp_jZmhiZzYWaxEvIL-tqebMzo/edit#heading=h.8tiycl1h7rns for details.
+func checkIfOrphanedGitOpsDeploymentSyncRun(ctx context.Context, event eventlooptypes.EventLoopMessage,
+	orphanedResources map[string]map[string]eventlooptypes.EventLoopEvent, log logr.Logger) string {
 
 	if event.Event.ReqResource == eventlooptypes.GitOpsDeploymentSyncRunTypeName {
 
+		// 1) Retrieve the SyncRun
 		syncRunCR := &v1alpha1.GitOpsDeploymentSyncRun{
 			ObjectMeta: metav1.ObjectMeta{
 				Name:      event.Event.Request.Name,
@@ -281,28 +311,57 @@ func handleOrphaned(ctx context.Context, event eventlooptypes.EventLoopMessage, 
 		}
 		if err := event.Event.Client.Get(ctx, client.ObjectKeyFromObject(syncRunCR), syncRunCR); err != nil {
 			if apierr.IsNotFound(err) {
-				log.V(sharedutil.LogLevel_Debug).Info("skipping orphaned resource that could no longer be found:", "resource", syncRunCR.ObjectMeta)
-				return
+				log.V(sharedutil.LogLevel_Debug).Info("skipping potentially orphaned resource that could no longer be found:", "resource", syncRunCR.ObjectMeta)
+				return ""
 			} else {
-				log.Error(err, "unexpected client error on retrieving orphaned syncrun object", "resource", syncRunCR.ObjectMeta)
-				return
+				log.Error(err, "unexpected client error on retrieving syncrun object", "resource", syncRunCR.ObjectMeta)
+				return ""
 			}
 		}
 
-		// TODO: GITOPSRVCE-224: Make sure it's not still orphaned before adding it to orphanedResources.
+		// 2) Retrieve the GitOpsDeployment, referenced by the SyncRun, to see if the SyncRun is orphaned
+		gitopsDeplCR := &v1alpha1.GitOpsDeployment{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      syncRunCR.Spec.GitopsDeploymentName,
+				Namespace: event.Event.Request.Namespace,
+			},
+		}
+		if err := event.Event.Client.Get(ctx, client.ObjectKeyFromObject(gitopsDeplCR), gitopsDeplCR); err != nil {
 
+			if !apierr.IsNotFound(err) {
+				log.Error(err, "unexpected client error on retrieving GitOpsDeployment references by GitOpsDeploymentSyncRun", "resource", syncRunCR.ObjectMeta)
+				return ""
+			}
+
+			log.V(sharedutil.LogLevel_Debug).Info("was unable to locate GitOpsDeployment referenced by GitOpsDeploymentSyncRun, so the SyncRun is still orphaned.")
+
+			// Otherwise, the GitOpsDeployment couldn't be found: therefore the SyncRun is orphaned, so continue.
+
+		} else {
+			// The resource is not orphaned, so our work is done: return the GitOpsDeployment referenced by the GitOpsDeploymentSyncRun
+			return gitopsDeplCR.Name
+		}
+
+		// At this point, the SyncRun exists, but the GitOpsDeployment referenced by it doesn't
+
+		// 3) Retrieve the list of orphaned resources that are depending on the GitOpsDeployment name referenced in the spec
+		// field of the SyncRun.
 		gitopsDeplMap, exists := orphanedResources[syncRunCR.Spec.GitopsDeploymentName]
 		if !exists {
 			gitopsDeplMap = map[string]eventlooptypes.EventLoopEvent{}
 			orphanedResources[syncRunCR.Spec.GitopsDeploymentName] = gitopsDeplMap
 		}
 
+		// 4) Add this resource event to that list
 		log.V(sharedutil.LogLevel_Debug).Info("Adding syncrun CR to orphaned resources list, name: " + syncRunCR.Name + ", missing gitopsdepl name: " + syncRunCR.Spec.GitopsDeploymentName)
 		gitopsDeplMap[syncRunCR.Name] = *event.Event
 
+		return ""
+
 	} else {
+
 		log.Error(nil, "SEVERE: unexpected event resource type in handleOrphaned")
-		return
+		return ""
 	}
 
 }
@@ -321,6 +380,7 @@ func unorphanResourcesIfPossible(ctx context.Context, event eventlooptypes.Event
 				Namespace: event.Event.Request.Namespace,
 			},
 		}
+
 		if err := event.Event.Client.Get(ctx, client.ObjectKeyFromObject(gitopsDeplCR), gitopsDeplCR); err != nil {
 			// log as warning, but continue.
 			log.V(sharedutil.LogLevel_Warn).Error(err, "unexpected client error on retrieving gitopsdepl")
@@ -331,9 +391,6 @@ func unorphanResourcesIfPossible(ctx context.Context, event eventlooptypes.Event
 			for index := range gitopsDeplMap {
 
 				orphanedResourceEvent := gitopsDeplMap[index]
-
-				// Unorphan the resource
-				orphanedResourceEvent.AssociatedGitopsDeplUID = string(gitopsDeplCR.UID)
 
 				requeueEvents = append(requeueEvents, eventlooptypes.EventLoopMessage{
 					MessageType: eventlooptypes.ApplicationEventLoopMessageType_Event,
@@ -348,7 +405,8 @@ func unorphanResourcesIfPossible(ctx context.Context, event eventlooptypes.Event
 			// The orphaned events will be reprocessed after the gitopsdepl is processed.
 			go func() {
 				for _, eventToRequeue := range requeueEvents {
-					log.V(sharedutil.LogLevel_Debug).Info("requeueing orphaned resource: " + eventToRequeue.Event.Request.Name + ", for parent: " + eventToRequeue.Event.AssociatedGitopsDeplUID)
+					log.V(sharedutil.LogLevel_Debug).Info("requeueing orphaned resource: " + eventToRequeue.Event.Request.Name +
+						", for parent: " + gitopsDeplCR.Name + " in namespace " + gitopsDeplCR.Namespace)
 					input <- workspaceEventLoopMessage{
 						messageType: workspaceEventLoopMessageType_Event,
 						payload:     eventToRequeue,
