@@ -1,8 +1,10 @@
 package eventloop
 
 import (
+	"context"
 	"fmt"
 	"sync"
+	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -22,7 +24,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
 
-// This test scenario requires, tests to run in order, hence using 'Ordered' decorator.
+// This test scenario requires tests to run in order, hence using 'Ordered' decorator.
 var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 
 	Context("Workspace event loop responds to channel events", func() {
@@ -35,6 +37,8 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 		var tAELF *testApplicationEventLoopFactory
 		var msgTemp eventlooptypes.EventLoopMessage
 		var workspaceEventLoopRouter WorkspaceEventLoopRouterStruct
+
+		var k8sClient client.WithWatch
 
 		// Start the workspace event loop using single ApplicationEventLoopFactory object,
 		// this way all tests can keep track of number of event loops created by other tests.
@@ -53,9 +57,15 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 
 			// Start the workspace event loop with our custom test factory, so that we can capture output
 			workspaceEventLoopRouter = newWorkspaceEventLoopRouterWithFactory(string(apiNamespace.UID), tAELF)
+
+			k8sClient = fake.NewClientBuilder().
+				WithScheme(scheme).
+				WithObjects(apiNamespace, argocdNamespace, kubesystemNamespace).
+				Build()
+
 		})
 
-		It("Should pass events received on input channel to application event loop", func() {
+		It("should pass events received on input channel to application event loop", func() {
 
 			gitopsDepl := &managedgitopsv1alpha1.GitOpsDeployment{
 				ObjectMeta: metav1.ObjectMeta{
@@ -69,10 +79,8 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 				},
 			}
 
-			k8sClient := fake.NewClientBuilder().
-				WithScheme(scheme).
-				WithObjects(gitopsDepl, apiNamespace, argocdNamespace, kubesystemNamespace).
-				Build()
+			err := k8sClient.Create(context.Background(), gitopsDepl)
+			Expect(err).To(BeNil())
 
 			// Simulate a GitOpsDeployment modified event
 			msg := eventlooptypes.EventLoopMessage{
@@ -82,13 +90,12 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 					Request: reconcile.Request{
 						NamespacedName: types.NamespacedName{
 							Namespace: apiNamespace.Name,
-							Name:      apiNamespace.Name,
+							Name:      gitopsDepl.Name,
 						},
 					},
-					Client:                  k8sClient,
-					ReqResource:             eventlooptypes.GitOpsDeploymentTypeName,
-					AssociatedGitopsDeplUID: string(gitopsDepl.UID),
-					WorkspaceID:             string(apiNamespace.UID),
+					Client:      k8sClient,
+					ReqResource: eventlooptypes.GitOpsDeploymentTypeName,
+					WorkspaceID: string(apiNamespace.UID),
 				},
 			}
 
@@ -102,7 +109,8 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 			Expect(tAELF.numberOfEventLoopsCreated).To(Equal(1))
 		})
 
-		It("Should not create new application event loop when 2nd event is passed with same ID", func() {
+		It("should not create new application event loop when 2nd event is passed with same name/ID", func() {
+
 			gitopsDepl := &managedgitopsv1alpha1.GitOpsDeployment{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "my-gitops-depl",
@@ -115,30 +123,24 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 				},
 			}
 
-			k8sClient := fake.NewClientBuilder().
-				WithScheme(scheme).
-				WithObjects(gitopsDepl, apiNamespace, argocdNamespace, kubesystemNamespace).
-				Build()
-
-			// Simulate a GitOpsDeployment modified event
+			By("simulating a GitOpsDeployment modified event")
 			msg := eventlooptypes.EventLoopMessage{
 				MessageType: eventlooptypes.ApplicationEventLoopMessageType_Event,
 				Event: &eventlooptypes.EventLoopEvent{
 					EventType: eventlooptypes.DeploymentModified,
 					Request: reconcile.Request{
 						NamespacedName: types.NamespacedName{
-							Namespace: apiNamespace.Name,
-							Name:      apiNamespace.Name,
+							Namespace: gitopsDepl.Namespace,
+							Name:      gitopsDepl.Name,
 						},
 					},
-					Client:                  k8sClient,
-					ReqResource:             eventlooptypes.GitOpsDeploymentTypeName,
-					AssociatedGitopsDeplUID: string(gitopsDepl.UID),
-					WorkspaceID:             string(apiNamespace.UID),
+					Client:      k8sClient,
+					ReqResource: eventlooptypes.GitOpsDeploymentTypeName,
+					WorkspaceID: string(apiNamespace.UID),
 				},
 			}
 
-			// Sending an event to the workspace event loop should cause an event to be sent on the output channel
+			By("sending an event to the workspace event loop should cause an event to be sent on the output channel")
 			workspaceEventLoopRouter.SendMessage(msg)
 			result := <-tAELF.outputChannel
 
@@ -149,7 +151,7 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 
 		})
 
-		It("Should create new application event loop when 3rd event is passed with diferent ID", func() {
+		It("should not create new application event loop when 3rd event is passed with same name but different UID", func() {
 			gitopsDepl := &managedgitopsv1alpha1.GitOpsDeployment{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "my-gitops-depl",
@@ -162,12 +164,15 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 				},
 			}
 
-			k8sClient := fake.NewClientBuilder().
-				WithScheme(scheme).
-				WithObjects(gitopsDepl, apiNamespace, argocdNamespace, kubesystemNamespace).
-				Build()
+			By("deleting the old GitOpsDeployment from the previous test")
+			err := k8sClient.Delete(context.Background(), gitopsDepl)
+			Expect(err).To(BeNil())
 
-			// Simulate a GitOpsDeployment modified event
+			By("creating a new GitOpsDeployment with the same name as the previous one, but a different UID")
+			err = k8sClient.Create(context.Background(), gitopsDepl)
+			Expect(err).To(BeNil())
+
+			By("simulating a GitOpsDeployment modified event")
 			msg := eventlooptypes.EventLoopMessage{
 				MessageType: eventlooptypes.ApplicationEventLoopMessageType_Event,
 				Event: &eventlooptypes.EventLoopEvent{
@@ -175,28 +180,27 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 					Request: reconcile.Request{
 						NamespacedName: types.NamespacedName{
 							Namespace: apiNamespace.Name,
-							Name:      apiNamespace.Name,
+							Name:      gitopsDepl.Name,
 						},
 					},
-					Client:                  k8sClient,
-					ReqResource:             eventlooptypes.GitOpsDeploymentTypeName,
-					AssociatedGitopsDeplUID: string(gitopsDepl.UID),
-					WorkspaceID:             string(apiNamespace.UID),
+					Client:      k8sClient,
+					ReqResource: eventlooptypes.GitOpsDeploymentTypeName,
+					WorkspaceID: string(apiNamespace.UID),
 				},
 			}
 
-			// Sending an event to the workspace event loop should cause an event to be sent on the output channel
+			By("sending an event to the workspace event loop should cause an event to be sent on the output channel")
 			workspaceEventLoopRouter.SendMessage(msg)
 			result := <-tAELF.outputChannel
 
 			Expect(result).NotTo(BeNil())
 			Expect(result.Event).To(Equal(msg.Event))
 			Expect(result.MessageType).To(Equal(msg.MessageType))
-			Expect(tAELF.numberOfEventLoopsCreated).To(Equal(2))
+			Expect(tAELF.numberOfEventLoopsCreated).To(Equal(1))
 
 		})
 
-		It("Should not pass an orphaned event to application event loop.", func() {
+		It("should not pass an orphaned event to application event loop.", func() {
 
 			gitopsDeplSyncRun := &managedgitopsv1alpha1.GitOpsDeploymentSyncRun{
 				ObjectMeta: metav1.ObjectMeta{
@@ -210,12 +214,11 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 				},
 			}
 
-			k8sClient := fake.NewClientBuilder().
-				WithScheme(scheme).
-				WithObjects(gitopsDeplSyncRun, apiNamespace, argocdNamespace, kubesystemNamespace).
-				Build()
+			By("creating a new orphaned GitOpsDeploymentSync")
+			err = k8sClient.Create(context.Background(), gitopsDeplSyncRun)
+			Expect(err).To(BeNil())
 
-			// Simulate a GitOpsDeploymentSyncRun event
+			By("simulating a GitOpsDeploymentSyncRun event")
 			msg := eventlooptypes.EventLoopMessage{
 				MessageType: eventlooptypes.ApplicationEventLoopMessageType_Event,
 				Event: &eventlooptypes.EventLoopEvent{
@@ -226,10 +229,9 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 							Name:      "my-gitops-depl-sync",
 						},
 					},
-					Client:                  k8sClient,
-					ReqResource:             eventlooptypes.GitOpsDeploymentSyncRunTypeName,
-					AssociatedGitopsDeplUID: "orphaned",
-					WorkspaceID:             string(apiNamespace.UID),
+					Client:      k8sClient,
+					ReqResource: eventlooptypes.GitOpsDeploymentSyncRunTypeName,
+					WorkspaceID: string(apiNamespace.UID),
 				},
 			}
 
@@ -269,12 +271,11 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 				},
 			}
 
-			k8sClient := fake.NewClientBuilder().
-				WithScheme(scheme).
-				WithObjects(gitopsDepl, apiNamespace, argocdNamespace, kubesystemNamespace).
-				Build()
+			By("creating a new GitOpsDeployment, to unorphan the GitOpsDeploymentSync")
+			err = k8sClient.Create(context.Background(), gitopsDepl)
+			Expect(err).To(BeNil())
 
-			// Simulate a GitOpsDeployment modified event
+			By("simulating a GitOpsDeployment modified event")
 			msg := eventlooptypes.EventLoopMessage{
 				MessageType: eventlooptypes.ApplicationEventLoopMessageType_Event,
 				Event: &eventlooptypes.EventLoopEvent{
@@ -285,25 +286,24 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 							Name:      gitopsDepl.Name,
 						},
 					},
-					Client:                  k8sClient,
-					ReqResource:             eventlooptypes.GitOpsDeploymentTypeName,
-					AssociatedGitopsDeplUID: string(gitopsDepl.UID),
-					WorkspaceID:             string(apiNamespace.UID),
+					Client:      k8sClient,
+					ReqResource: eventlooptypes.GitOpsDeploymentTypeName,
+					WorkspaceID: string(apiNamespace.UID),
 				},
 			}
 
-			// Create goroutine and pass event,
-			//because loop in workspaceEventLoopRouter will keep on waiting for new event to be received and test will get stuck here.
+			// Create goroutine and pass event, because loop in workspaceEventLoopRouter will keep on waiting for new event
+			// to be received and test will get stuck here.
 			go func() {
 				workspaceEventLoopRouter.SendMessage(msg)
 			}()
 
 			Eventually(func() bool {
-				// We use a function here to check if number is 3
-				return tAELF.numberOfEventLoopsCreated == 3
-			}).Should(BeTrue())
+				// We use a function here to check if number is 2
+				return tAELF.numberOfEventLoopsCreated == 2
+			}, time.Second*240).Should(BeTrue())
 
-			By("Read the events from the output channel, make sure they are the ones we expect, now that the gitopsdeploymentsyncrun in unorphaned")
+			By("reading the events from the output channel, make sure they are the ones we expect, now that the gitopsdeploymentsyncrun in unorphaned")
 
 			deploymentModifiedMsg := <-tAELF.outputChannel
 
@@ -361,7 +361,7 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 					},
 				}
 
-				tAELF.outputChannel[string(gitopsDepl.UID)] = make(chan eventlooptypes.EventLoopMessage)
+				tAELF.outputChannel2[gitopsDepl.Name] = make(chan eventlooptypes.EventLoopMessage)
 
 				res = append(res, *gitopsDepl)
 
@@ -373,21 +373,20 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 						Request: reconcile.Request{
 							NamespacedName: types.NamespacedName{
 								Namespace: apiNamespace.Name,
-								Name:      apiNamespace.Name,
+								Name:      gitopsDepl.Name,
 							},
 						},
-						Client:                  k8sClient,
-						ReqResource:             eventlooptypes.GitOpsDeploymentTypeName,
-						AssociatedGitopsDeplUID: string(gitopsDepl.UID),
-						WorkspaceID:             string(apiNamespace.UID),
+						Client:      k8sClient,
+						ReqResource: eventlooptypes.GitOpsDeploymentTypeName,
+						WorkspaceID: string(apiNamespace.UID),
 					},
 				}
 
-				By("Sending an event to the workspace event loop should cause an event to be sent on the output channel")
+				By("sending an event to the workspace event loop should cause an event to be seen on the output channel")
 				workspaceEventLoopRouter.SendMessage(msg)
 
-				By("waiting for the respond to be received")
-				result := <-tAELF.outputChannel[string(gitopsDepl.UID)]
+				By("waiting for the responce to be received")
+				result := <-tAELF.outputChannel2[gitopsDepl.Name]
 				Expect(result).NotTo(BeNil())
 				Expect(result.Event).To(Equal(msg.Event))
 				Expect(result.MessageType).To(Equal(msg.MessageType))
@@ -401,7 +400,7 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 
 			By("Starting the workspace event loop with our custom test factory, so that we can capture output")
 			tAELF := &managedEnvironmentTestApplicationEventLoopFactory{
-				outputChannel: map[string]chan eventlooptypes.EventLoopMessage{},
+				outputChannel2: map[string]chan eventlooptypes.EventLoopMessage{},
 			}
 			workspaceEventLoopRouter := newWorkspaceEventLoopRouterWithFactory(string(apiNamespace.UID), tAELF)
 
@@ -435,7 +434,7 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 
 			By("Starting the workspace event loop with our custom test factory, so that we can capture output")
 			tAELF := &managedEnvironmentTestApplicationEventLoopFactory{
-				outputChannel: map[string]chan eventlooptypes.EventLoopMessage{},
+				outputChannel2: map[string]chan eventlooptypes.EventLoopMessage{},
 			}
 			workspaceEventLoopRouter := newWorkspaceEventLoopRouterWithFactory(string(apiNamespace.UID), tAELF)
 
@@ -444,7 +443,8 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 				WithObjects( /*gitopsDepl,*/ apiNamespace, argocdNamespace, kubesystemNamespace).
 				Build()
 
-			By("creating simulating GitOpsDeployment events, causing the workspace event router to start an application event loop for them")
+			By("creating simulating GitOpsDeployment events, causing the workspace event router to start an application event loop " +
+				"for them")
 			gitopsDeployments := simulateGitOpsDeployments(numGitOpsDeploymentsToSimulate, tAELF, &workspaceEventLoopRouter, k8sClient)
 
 			Expect(tAELF.numberOfEventLoopsCreated).To(Equal(numGitOpsDeploymentsToSimulate),
@@ -478,7 +478,7 @@ var _ = Describe("Workspace Event Loop Test", Ordered, func() {
 				By("starting a goroutine which writes all received events to eventLoopMsgsReceived for " + string(gitopsDeployment.UID))
 				go func() {
 					for {
-						fromOutputChan := <-tAELF.outputChannel[string(gitopsDeployment.UID)]
+						fromOutputChan := <-tAELF.outputChannel2[string(gitopsDeployment.Name)]
 						GinkgoWriter.Println("event received on gofunc", string(gitopsDeployment.UID), fromOutputChan)
 						mutex.Lock()
 						eventLoopMsgsReceived = append(eventLoopMsgsReceived, fromOutputChan)
@@ -538,7 +538,7 @@ var _ applicationEventQueueLoopFactory = &testApplicationEventLoopFactory{}
 
 // Instead of starting a new application event queue loop (like the default implementation of applictionEventQueueLoopFactory)
 // we instead just return a previously provided channel.
-func (ta *testApplicationEventLoopFactory) startApplicationEventQueueLoop(gitopsDeplID string, workspaceID string,
+func (ta *testApplicationEventLoopFactory) startApplicationEventQueueLoop(ctx context.Context, gitopsDeplName string, gitopsDeploymentNamespace string, workspaceID string,
 	sharedResourceEventLoop *shared_resource_loop.SharedResourceEventLoop) chan eventlooptypes.EventLoopMessage {
 
 	ta.mutex.Lock()
@@ -551,7 +551,7 @@ func (ta *testApplicationEventLoopFactory) startApplicationEventQueueLoop(gitops
 
 // testApplicationEventLoopFactory is a mock applicationEventQueueLoopFactory, used for unit tests above.
 type managedEnvironmentTestApplicationEventLoopFactory struct {
-	outputChannel             map[string]chan eventlooptypes.EventLoopMessage
+	outputChannel2            map[string]chan eventlooptypes.EventLoopMessage
 	numberOfEventLoopsCreated int
 	mutex                     sync.Mutex
 }
@@ -560,13 +560,13 @@ var _ applicationEventQueueLoopFactory = &managedEnvironmentTestApplicationEvent
 
 // Instead of starting a new application event queue loop (like the default implementation of applictionEventQueueLoopFactory)
 // we instead just return a previously provided channel.
-func (ta *managedEnvironmentTestApplicationEventLoopFactory) startApplicationEventQueueLoop(gitopsDeplID string, workspaceID string,
-	sharedResourceEventLoop *shared_resource_loop.SharedResourceEventLoop) chan eventlooptypes.EventLoopMessage {
+func (ta *managedEnvironmentTestApplicationEventLoopFactory) startApplicationEventQueueLoop(ctx context.Context, gitopsDeplName string,
+	gitopsDeplNamespace string, workspaceID string, sharedResourceEventLoop *shared_resource_loop.SharedResourceEventLoop) chan eventlooptypes.EventLoopMessage {
 
 	ta.mutex.Lock()
 	defer ta.mutex.Unlock()
 
 	// Increase count by 1 if new application event loop is created
 	ta.numberOfEventLoopsCreated++
-	return ta.outputChannel[gitopsDeplID]
+	return ta.outputChannel2[gitopsDeplName]
 }
