@@ -1,4 +1,4 @@
-// Code generated by mockery v2.10.0. DO NOT EDIT.
+// Code generated by mockery v2.14.1. DO NOT EDIT.
 
 package mocks
 
@@ -24,11 +24,11 @@ type ApplicationServiceClient struct {
 
 // Create provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) Create(ctx context.Context, in *application.ApplicationCreateRequest, opts ...grpc.CallOption) (*v1alpha1.Application, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -54,11 +54,11 @@ func (_m *ApplicationServiceClient) Create(ctx context.Context, in *application.
 
 // Delete provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) Delete(ctx context.Context, in *application.ApplicationDeleteRequest, opts ...grpc.CallOption) (*application.ApplicationResponse, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -84,11 +84,11 @@ func (_m *ApplicationServiceClient) Delete(ctx context.Context, in *application.
 
 // DeleteResource provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) DeleteResource(ctx context.Context, in *application.ApplicationResourceDeleteRequest, opts ...grpc.CallOption) (*application.ApplicationResponse, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -114,11 +114,11 @@ func (_m *ApplicationServiceClient) DeleteResource(ctx context.Context, in *appl
 
 // Get provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) Get(ctx context.Context, in *application.ApplicationQuery, opts ...grpc.CallOption) (*v1alpha1.Application, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -144,11 +144,11 @@ func (_m *ApplicationServiceClient) Get(ctx context.Context, in *application.App
 
 // GetApplicationSyncWindows provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) GetApplicationSyncWindows(ctx context.Context, in *application.ApplicationSyncWindowsQuery, opts ...grpc.CallOption) (*application.ApplicationSyncWindowsResponse, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -174,11 +174,11 @@ func (_m *ApplicationServiceClient) GetApplicationSyncWindows(ctx context.Contex
 
 // GetManifests provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) GetManifests(ctx context.Context, in *application.ApplicationManifestQuery, opts ...grpc.CallOption) (*apiclient.ManifestResponse, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -202,13 +202,43 @@ func (_m *ApplicationServiceClient) GetManifests(ctx context.Context, in *applic
 	return r0, r1
 }
 
-// GetResource provides a mock function with given fields: ctx, in, opts
-func (_m *ApplicationServiceClient) GetResource(ctx context.Context, in *application.ApplicationResourceRequest, opts ...grpc.CallOption) (*application.ApplicationResourceResponse, error) {
-	_va := make([]any, len(opts))
+// GetManifestsWithFiles provides a mock function with given fields: ctx, opts
+func (_m *ApplicationServiceClient) GetManifestsWithFiles(ctx context.Context, opts ...grpc.CallOption) (application.ApplicationService_GetManifestsWithFilesClient, error) {
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
+	_ca = append(_ca, ctx)
+	_ca = append(_ca, _va...)
+	ret := _m.Called(_ca...)
+
+	var r0 application.ApplicationService_GetManifestsWithFilesClient
+	if rf, ok := ret.Get(0).(func(context.Context, ...grpc.CallOption) application.ApplicationService_GetManifestsWithFilesClient); ok {
+		r0 = rf(ctx, opts...)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(application.ApplicationService_GetManifestsWithFilesClient)
+		}
+	}
+
+	var r1 error
+	if rf, ok := ret.Get(1).(func(context.Context, ...grpc.CallOption) error); ok {
+		r1 = rf(ctx, opts...)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
+// GetResource provides a mock function with given fields: ctx, in, opts
+func (_m *ApplicationServiceClient) GetResource(ctx context.Context, in *application.ApplicationResourceRequest, opts ...grpc.CallOption) (*application.ApplicationResourceResponse, error) {
+	_va := make([]interface{}, len(opts))
+	for _i := range opts {
+		_va[_i] = opts[_i]
+	}
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -234,11 +264,11 @@ func (_m *ApplicationServiceClient) GetResource(ctx context.Context, in *applica
 
 // List provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) List(ctx context.Context, in *application.ApplicationQuery, opts ...grpc.CallOption) (*v1alpha1.ApplicationList, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -264,11 +294,11 @@ func (_m *ApplicationServiceClient) List(ctx context.Context, in *application.Ap
 
 // ListResourceActions provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) ListResourceActions(ctx context.Context, in *application.ApplicationResourceRequest, opts ...grpc.CallOption) (*application.ResourceActionsListResponse, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -294,11 +324,11 @@ func (_m *ApplicationServiceClient) ListResourceActions(ctx context.Context, in 
 
 // ListResourceEvents provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) ListResourceEvents(ctx context.Context, in *application.ApplicationResourceEventsQuery, opts ...grpc.CallOption) (*v1.EventList, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -324,11 +354,11 @@ func (_m *ApplicationServiceClient) ListResourceEvents(ctx context.Context, in *
 
 // ManagedResources provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) ManagedResources(ctx context.Context, in *application.ResourcesQuery, opts ...grpc.CallOption) (*application.ManagedResourcesResponse, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -354,11 +384,11 @@ func (_m *ApplicationServiceClient) ManagedResources(ctx context.Context, in *ap
 
 // Patch provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) Patch(ctx context.Context, in *application.ApplicationPatchRequest, opts ...grpc.CallOption) (*v1alpha1.Application, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -384,11 +414,11 @@ func (_m *ApplicationServiceClient) Patch(ctx context.Context, in *application.A
 
 // PatchResource provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) PatchResource(ctx context.Context, in *application.ApplicationResourcePatchRequest, opts ...grpc.CallOption) (*application.ApplicationResourceResponse, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -414,11 +444,11 @@ func (_m *ApplicationServiceClient) PatchResource(ctx context.Context, in *appli
 
 // PodLogs provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) PodLogs(ctx context.Context, in *application.ApplicationPodLogsQuery, opts ...grpc.CallOption) (application.ApplicationService_PodLogsClient, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -444,11 +474,11 @@ func (_m *ApplicationServiceClient) PodLogs(ctx context.Context, in *application
 
 // ResourceTree provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) ResourceTree(ctx context.Context, in *application.ResourcesQuery, opts ...grpc.CallOption) (*v1alpha1.ApplicationTree, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -474,11 +504,11 @@ func (_m *ApplicationServiceClient) ResourceTree(ctx context.Context, in *applic
 
 // RevisionMetadata provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) RevisionMetadata(ctx context.Context, in *application.RevisionMetadataQuery, opts ...grpc.CallOption) (*v1alpha1.RevisionMetadata, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -504,11 +534,11 @@ func (_m *ApplicationServiceClient) RevisionMetadata(ctx context.Context, in *ap
 
 // Rollback provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) Rollback(ctx context.Context, in *application.ApplicationRollbackRequest, opts ...grpc.CallOption) (*v1alpha1.Application, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -534,11 +564,11 @@ func (_m *ApplicationServiceClient) Rollback(ctx context.Context, in *applicatio
 
 // RunResourceAction provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) RunResourceAction(ctx context.Context, in *application.ResourceActionRunRequest, opts ...grpc.CallOption) (*application.ApplicationResponse, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -564,11 +594,11 @@ func (_m *ApplicationServiceClient) RunResourceAction(ctx context.Context, in *a
 
 // Sync provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) Sync(ctx context.Context, in *application.ApplicationSyncRequest, opts ...grpc.CallOption) (*v1alpha1.Application, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -594,11 +624,11 @@ func (_m *ApplicationServiceClient) Sync(ctx context.Context, in *application.Ap
 
 // TerminateOperation provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) TerminateOperation(ctx context.Context, in *application.OperationTerminateRequest, opts ...grpc.CallOption) (*application.OperationTerminateResponse, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -624,11 +654,11 @@ func (_m *ApplicationServiceClient) TerminateOperation(ctx context.Context, in *
 
 // Update provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) Update(ctx context.Context, in *application.ApplicationUpdateRequest, opts ...grpc.CallOption) (*v1alpha1.Application, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -654,11 +684,11 @@ func (_m *ApplicationServiceClient) Update(ctx context.Context, in *application.
 
 // UpdateSpec provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) UpdateSpec(ctx context.Context, in *application.ApplicationUpdateSpecRequest, opts ...grpc.CallOption) (*v1alpha1.ApplicationSpec, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -684,11 +714,11 @@ func (_m *ApplicationServiceClient) UpdateSpec(ctx context.Context, in *applicat
 
 // Watch provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) Watch(ctx context.Context, in *application.ApplicationQuery, opts ...grpc.CallOption) (application.ApplicationService_WatchClient, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -714,11 +744,11 @@ func (_m *ApplicationServiceClient) Watch(ctx context.Context, in *application.A
 
 // WatchResourceTree provides a mock function with given fields: ctx, in, opts
 func (_m *ApplicationServiceClient) WatchResourceTree(ctx context.Context, in *application.ResourcesQuery, opts ...grpc.CallOption) (application.ApplicationService_WatchResourceTreeClient, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -740,4 +770,19 @@ func (_m *ApplicationServiceClient) WatchResourceTree(ctx context.Context, in *a
 	}
 
 	return r0, r1
+}
+
+type mockConstructorTestingTNewApplicationServiceClient interface {
+	mock.TestingT
+	Cleanup(func())
+}
+
+// NewApplicationServiceClient creates a new instance of ApplicationServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+func NewApplicationServiceClient(t mockConstructorTestingTNewApplicationServiceClient) *ApplicationServiceClient {
+	mock := &ApplicationServiceClient{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
 }
