@@ -1,4 +1,4 @@
-// Code generated by mockery v2.10.0. DO NOT EDIT.
+// Code generated by mockery v2.14.1. DO NOT EDIT.
 
 package mocks
 
@@ -7,6 +7,8 @@ import (
 	account "github.com/argoproj/argo-cd/v2/pkg/apiclient/account"
 
 	application "github.com/argoproj/argo-cd/v2/pkg/apiclient/application"
+
+	applicationset "github.com/argoproj/argo-cd/v2/pkg/apiclient/applicationset"
 
 	certificate "github.com/argoproj/argo-cd/v2/pkg/apiclient/certificate"
 
@@ -21,6 +23,8 @@ import (
 	io "io"
 
 	mock "github.com/stretchr/testify/mock"
+
+	notification "github.com/argoproj/argo-cd/v2/pkg/apiclient/notification"
 
 	oauth2 "golang.org/x/oauth2"
 
@@ -197,6 +201,63 @@ func (_m *Client) NewApplicationClientOrDie() (io.Closer, application.Applicatio
 	return r0, r1
 }
 
+// NewApplicationSetClient provides a mock function with given fields:
+func (_m *Client) NewApplicationSetClient() (io.Closer, applicationset.ApplicationSetServiceClient, error) {
+	ret := _m.Called()
+
+	var r0 io.Closer
+	if rf, ok := ret.Get(0).(func() io.Closer); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(io.Closer)
+		}
+	}
+
+	var r1 applicationset.ApplicationSetServiceClient
+	if rf, ok := ret.Get(1).(func() applicationset.ApplicationSetServiceClient); ok {
+		r1 = rf()
+	} else {
+		if ret.Get(1) != nil {
+			r1 = ret.Get(1).(applicationset.ApplicationSetServiceClient)
+		}
+	}
+
+	var r2 error
+	if rf, ok := ret.Get(2).(func() error); ok {
+		r2 = rf()
+	} else {
+		r2 = ret.Error(2)
+	}
+
+	return r0, r1, r2
+}
+
+// NewApplicationSetClientOrDie provides a mock function with given fields:
+func (_m *Client) NewApplicationSetClientOrDie() (io.Closer, applicationset.ApplicationSetServiceClient) {
+	ret := _m.Called()
+
+	var r0 io.Closer
+	if rf, ok := ret.Get(0).(func() io.Closer); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(io.Closer)
+		}
+	}
+
+	var r1 applicationset.ApplicationSetServiceClient
+	if rf, ok := ret.Get(1).(func() applicationset.ApplicationSetServiceClient); ok {
+		r1 = rf()
+	} else {
+		if ret.Get(1) != nil {
+			r1 = ret.Get(1).(applicationset.ApplicationSetServiceClient)
+		}
+	}
+
+	return r0, r1
+}
+
 // NewCertClient provides a mock function with given fields:
 func (_m *Client) NewCertClient() (io.Closer, certificate.CertificateServiceClient, error) {
 	ret := _m.Called()
@@ -362,6 +423,63 @@ func (_m *Client) NewGPGKeyClientOrDie() (io.Closer, gpgkey.GPGKeyServiceClient)
 	} else {
 		if ret.Get(1) != nil {
 			r1 = ret.Get(1).(gpgkey.GPGKeyServiceClient)
+		}
+	}
+
+	return r0, r1
+}
+
+// NewNotificationClient provides a mock function with given fields:
+func (_m *Client) NewNotificationClient() (io.Closer, notification.NotificationServiceClient, error) {
+	ret := _m.Called()
+
+	var r0 io.Closer
+	if rf, ok := ret.Get(0).(func() io.Closer); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(io.Closer)
+		}
+	}
+
+	var r1 notification.NotificationServiceClient
+	if rf, ok := ret.Get(1).(func() notification.NotificationServiceClient); ok {
+		r1 = rf()
+	} else {
+		if ret.Get(1) != nil {
+			r1 = ret.Get(1).(notification.NotificationServiceClient)
+		}
+	}
+
+	var r2 error
+	if rf, ok := ret.Get(2).(func() error); ok {
+		r2 = rf()
+	} else {
+		r2 = ret.Error(2)
+	}
+
+	return r0, r1, r2
+}
+
+// NewNotificationClientOrDie provides a mock function with given fields:
+func (_m *Client) NewNotificationClientOrDie() (io.Closer, notification.NotificationServiceClient) {
+	ret := _m.Called()
+
+	var r0 io.Closer
+	if rf, ok := ret.Get(0).(func() io.Closer); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(io.Closer)
+		}
+	}
+
+	var r1 notification.NotificationServiceClient
+	if rf, ok := ret.Get(1).(func() notification.NotificationServiceClient); ok {
+		r1 = rf()
+	} else {
+		if ret.Get(1) != nil {
+			r1 = ret.Get(1).(notification.NotificationServiceClient)
 		}
 	}
 
@@ -756,4 +874,19 @@ func (_m *Client) WatchApplicationWithRetry(ctx context.Context, appName string,
 	}
 
 	return r0
+}
+
+type mockConstructorTestingTNewClient interface {
+	mock.TestingT
+	Cleanup(func())
+}
+
+// NewClient creates a new instance of Client. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+func NewClient(t mockConstructorTestingTNewClient) *Client {
+	mock := &Client{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
 }
