@@ -1,4 +1,4 @@
-// Code generated by mockery v2.10.0. DO NOT EDIT.
+// Code generated by mockery v2.14.1. DO NOT EDIT.
 
 package mocks
 
@@ -19,11 +19,11 @@ type SettingsServiceClient struct {
 
 // Get provides a mock function with given fields: ctx, in, opts
 func (_m *SettingsServiceClient) Get(ctx context.Context, in *settings.SettingsQuery, opts ...grpc.CallOption) (*settings.Settings, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -45,4 +45,19 @@ func (_m *SettingsServiceClient) Get(ctx context.Context, in *settings.SettingsQ
 	}
 
 	return r0, r1
+}
+
+type mockConstructorTestingTNewSettingsServiceClient interface {
+	mock.TestingT
+	Cleanup(func())
+}
+
+// NewSettingsServiceClient creates a new instance of SettingsServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+func NewSettingsServiceClient(t mockConstructorTestingTNewSettingsServiceClient) *SettingsServiceClient {
+	mock := &SettingsServiceClient{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
 }
