@@ -1,4 +1,4 @@
-// Code generated by mockery v2.10.0. DO NOT EDIT.
+// Code generated by mockery v2.14.1. DO NOT EDIT.
 
 package mocks
 
@@ -20,11 +20,11 @@ type VersionServiceClient struct {
 
 // Version provides a mock function with given fields: ctx, in, opts
 func (_m *VersionServiceClient) Version(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*version.VersionMessage, error) {
-	_va := make([]any, len(opts))
+	_va := make([]interface{}, len(opts))
 	for _i := range opts {
 		_va[_i] = opts[_i]
 	}
-	var _ca []any
+	var _ca []interface{}
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
@@ -46,4 +46,19 @@ func (_m *VersionServiceClient) Version(ctx context.Context, in *emptypb.Empty, 
 	}
 
 	return r0, r1
+}
+
+type mockConstructorTestingTNewVersionServiceClient interface {
+	mock.TestingT
+	Cleanup(func())
+}
+
+// NewVersionServiceClient creates a new instance of VersionServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+func NewVersionServiceClient(t mockConstructorTestingTNewVersionServiceClient) *VersionServiceClient {
+	mock := &VersionServiceClient{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
 }
