@@ -26,7 +26,11 @@ var _ = Describe("Argo CD Application", func() {
 			gitOpsDeployment := buildGitOpsDeploymentResource("my-gitops-depl-automated",
 				"https://github.com/redhat-appstudio/gitops-repository-template", "environments/overlays/dev",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Automated)
-			err := k8s.Create(&gitOpsDeployment)
+
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
+			err = k8s.Create(&gitOpsDeployment, k8sClient)
 			Expect(err).To(Succeed())
 
 			By("GitOpsDeployment should have expected health and status")
