@@ -31,7 +31,12 @@ const (
 var _ = Describe("GitOpsDeployment E2E tests", func() {
 
 	Context("Create, Update and Delete a GitOpsDeployment ", func() {
-		k8sClient, err := fixture.GetKubeClient()
+
+		// this assumes that service is running on non aware kcp client
+		config, err := fixture.GetSystemKubeConfig()
+		Expect(err).To(BeNil())
+
+		k8sClient, err := fixture.GetKubeClient(config)
 		Expect(err).To(BeNil())
 		ctx := context.Background()
 
@@ -114,8 +119,8 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 
 			Eventually(func() bool {
 
-				k8sclient, err := fixture.GetKubeClient()
-				Expect(err).To(BeNil())
+				k8sclient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+				Expect(err).To(Succeed())
 
 				for _, resourceValue := range expectedResourceStatusList {
 					ns := typed.NamespacedName{
@@ -158,7 +163,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 			gitOpsDeploymentResource.Spec.Destination.Environment = ""
 			gitOpsDeploymentResource.Spec.Destination.Namespace = fixture.GitOpsServiceE2ENamespace
 
-			err := k8s.Create(&gitOpsDeploymentResource)
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
+			err = k8s.Create(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			Eventually(gitOpsDeploymentResource, ArgoCDReconcileWaitTime, "1s").Should(
@@ -202,7 +210,7 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 
 			By("deleting the GitOpsDeployment resource and waiting for the resources to be deleted")
 
-			err = k8s.Delete(&gitOpsDeploymentResource)
+			err = k8s.Delete(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			expectAllResourcesToBeDeleted(expectedResourceStatusList)
@@ -218,7 +226,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 			gitOpsDeploymentResource.Spec.Destination.Environment = ""
 			gitOpsDeploymentResource.Spec.Destination.Namespace = fixture.GitOpsServiceE2ENamespace
 
-			err := k8s.Create(&gitOpsDeploymentResource)
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
+			err = k8s.Create(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 			Eventually(gitOpsDeploymentResource, ArgoCDReconcileWaitTime, "1s").Should(
 				SatisfyAll(
@@ -232,7 +243,7 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 			gitOpsDeploymentResource.Spec.Source.RepoURL = repoURL
 			gitOpsDeploymentResource.Spec.Source.Path = "environments/overlays/dev"
 
-			err = k8s.Update(&gitOpsDeploymentResource)
+			err = k8s.Update(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			err = k8sClient.Get(ctx, client.ObjectKeyFromObject(&gitOpsDeploymentResource), &gitOpsDeploymentResource)
@@ -266,7 +277,7 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 				),
 			)
 
-			err = k8s.Delete(&gitOpsDeploymentResource)
+			err = k8s.Delete(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 		})
 
@@ -276,7 +287,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 				repoURL, "environments/overlays/dev",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Automated)
 
-			err := k8s.Create(&gitOpsDeploymentResource)
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
+			err = k8s.Create(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			Eventually(gitOpsDeploymentResource, ArgoCDReconcileWaitTime, "1s").Should(
@@ -290,7 +304,7 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 
 			gitOpsDeploymentResource.Spec.Source.Path = "environments/overlays/staging"
 
-			err = k8s.Update(&gitOpsDeploymentResource)
+			err = k8s.Update(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			err = k8sClient.Get(ctx, client.ObjectKeyFromObject(&gitOpsDeploymentResource), &gitOpsDeploymentResource)
@@ -304,7 +318,7 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 				),
 			)
 
-			err = k8s.Delete(&gitOpsDeploymentResource)
+			err = k8s.Delete(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 		})
 
@@ -312,19 +326,22 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 
 			Expect(fixture.EnsureCleanSlate()).To(Succeed())
 
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
 			By("creating a new GitOpsDeployment resource")
 			gitOpsDeploymentResource := buildTargetRevisionGitOpsDeploymentResource("gitops-depl-test",
 				"https://github.com/managed-gitops-test-data/deployment-permutations-a", "pathB", "branchA",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Automated)
 
-			err := k8s.Create(&gitOpsDeploymentResource)
+			err = k8s.Create(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			createResourceStatusList_deploymentPermutations := createResourceStatusListFunction_deploymentPermutations()
 
 			createResourceStatusList_deploymentPermutations = append(createResourceStatusList_deploymentPermutations,
 				getResourceStatusList_deploymentPermutations("deployment-permutations-a-brancha-pathb")...)
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			Eventually(gitOpsDeploymentResource, ArgoCDReconcileWaitTime, "1s").Should(
@@ -337,13 +354,13 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 
 			By("ensuring changes in repo URL deploys the resources that are a part of the latest modification")
 
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			gitOpsDeploymentResource.Spec.Source.RepoURL = "https://github.com/managed-gitops-test-data/deployment-permutations-b"
 
 			//updating the CR with changes in repoURL
-			err = k8s.Update(&gitOpsDeploymentResource)
+			err = k8s.Update(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			expectedResourceStatusList := createResourceStatusListFunction_deploymentPermutations()
@@ -360,10 +377,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 				),
 			)
 			By("delete the GitOpsDeployment resource")
-			err = k8s.Delete(&gitOpsDeploymentResource)
+			err = k8s.Delete(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).ToNot(Succeed())
 
 			expectAllResourcesToBeDeleted(expectedResourceStatusList)
@@ -374,6 +391,9 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 
 			Expect(fixture.EnsureCleanSlate()).To(Succeed())
 
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
 			By("creating a new GitOpsDeployment resource")
 			gitOpsDeploymentResource := buildTargetRevisionGitOpsDeploymentResource("gitops-depl-test",
 				"https://github.com/managed-gitops-test-data/deployment-permutations-a", "pathB", "branchA",
@@ -381,13 +401,13 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 			gitOpsDeploymentResource.Spec.Destination.Environment = ""
 			gitOpsDeploymentResource.Spec.Destination.Namespace = fixture.GitOpsServiceE2ENamespace
 
-			err := k8s.Create(&gitOpsDeploymentResource)
+			err = k8s.Create(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			createResourceStatusList := createResourceStatusListFunction_deploymentPermutations()
 			createResourceStatusList = append(createResourceStatusList,
 				getResourceStatusList_deploymentPermutations("deployment-permutations-a-brancha-pathb")...)
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			expectedReconciledStateField := managedgitopsv1alpha1.ReconciledState{
@@ -413,13 +433,13 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 
 			By("ensuring changes in repo URL deploys the resources that are a part of the latest modification")
 
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			gitOpsDeploymentResource.Spec.Source.Path = "pathC"
 
 			//updating the CR with changes in repoURL
-			err = k8s.Update(&gitOpsDeploymentResource)
+			err = k8s.Update(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			expectedResourceStatusList := createResourceStatusListFunction_deploymentPermutations()
@@ -448,10 +468,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 			)
 			By("delete the GitOpsDeployment resource")
 
-			err = k8s.Delete(&gitOpsDeploymentResource)
+			err = k8s.Delete(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).ToNot(Succeed())
 
 			expectAllResourcesToBeDeleted(expectedResourceStatusList)
@@ -467,12 +487,15 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 				"https://github.com/managed-gitops-test-data/deployment-permutations-a", "pathB", "branchA",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Automated)
 
-			err := k8s.Create(&gitOpsDeploymentResource)
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
+			err = k8s.Create(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			createResourceStatusList := createResourceStatusListFunction_deploymentPermutations()
 			createResourceStatusList = append(createResourceStatusList, getResourceStatusList_deploymentPermutations("deployment-permutations-a-brancha-pathb")...)
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			Eventually(gitOpsDeploymentResource, ArgoCDReconcileWaitTime, "1s").Should(
@@ -485,13 +508,13 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 
 			By("ensuring changes in repo URL deploys the resources that are a part of the latest modification")
 
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			gitOpsDeploymentResource.Spec.Source.TargetRevision = "branchB"
 
 			//updating the CR with changes in repoURL
-			err = k8s.Update(&gitOpsDeploymentResource)
+			err = k8s.Update(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			expectedResourceStatusList := createResourceStatusListFunction_deploymentPermutations()
@@ -507,10 +530,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 			)
 			By("deleting the GitOpsDeployment resource")
 
-			err = k8s.Delete(&gitOpsDeploymentResource)
+			err = k8s.Delete(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).ToNot(Succeed())
 
 			expectAllResourcesToBeDeleted(expectedResourceStatusList)
@@ -526,17 +549,20 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 				"https://github.com/redhat-appstudio/gitops-repository-template", "environments/overlays/dev", "xyz",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Automated)
 
-			err := k8s.Create(&gitOpsDeploymentResource)
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
+			err = k8s.Create(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			By("ensuring changes in target revision deploys the resources that are a part of the latest modification")
 
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 			gitOpsDeploymentResource.Spec.Source.TargetRevision = "HEAD"
 
 			//updating the CR with changes in target revision
-			err = k8s.Update(&gitOpsDeploymentResource)
+			err = k8s.Update(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			expectedResourceStatusList_GitOpsRepositoryTemplateRepo := []managedgitopsv1alpha1.ResourceStatus{
@@ -561,10 +587,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 			)
 			By("delete the GitOpsDeployment resource")
 
-			err = k8s.Delete(&gitOpsDeploymentResource)
+			err = k8s.Delete(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).ToNot(Succeed())
 
 			expectAllResourcesToBeDeleted(expectedResourceStatusList_GitOpsRepositoryTemplateRepo)
@@ -580,7 +606,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 				"https://github.com/redhat-appstudio/gitops-repository-template", "environments/overlays/dev",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Automated)
 
-			err = k8s.Create(&gitOpsDeploymentResource)
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
+			err = k8s.Create(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			By("populating the resource list")
@@ -607,10 +636,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 			)
 			By("delete the GitOpsDeployment resource")
 
-			err = k8s.Delete(&gitOpsDeploymentResource)
+			err = k8s.Delete(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
-			err = k8s.Get(&gitOpsDeploymentResource)
+			err = k8s.Get(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).ToNot(Succeed())
 
 			expectAllResourcesToBeDeleted(expectedResourceStatusList)
@@ -625,7 +654,10 @@ var _ = Describe("GitOpsDeployment E2E tests", func() {
 				"invalid-url", "path/path/path",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Automated)
 
-			err := k8s.Create(&gitOpsDeploymentResource)
+			k8sClient, err := fixture.GetE2ETestUserWorkspaceKubeClient()
+			Expect(err).To(Succeed())
+
+			err = k8s.Create(&gitOpsDeploymentResource, k8sClient)
 			Expect(err).To(Succeed())
 
 			Eventually(gitOpsDeploymentResource, ArgoCDReconcileWaitTime, "1s").ShouldNot(
